+++ packages/shopify-app-remix/src/server/authenticate/admin/helpers/redirect-with-exitiframe.ts
@@ -11,14 +11,18 @@ export function redirectWithExitIframe(
   const url = new URL(request.url);
 
   const queryParams = url.searchParams;
+
   const host = api.utils.sanitizeHost(queryParams.get('host')!);
 
   queryParams.set('shop', shop);
-  queryParams.set('exitIframe', `${config.auth.path}?shop=${shop}`);
+
+  let destination = `${config.auth.path}?shop=${shop}`;
 
   if (host) {
     queryParams.set('host', host);
+    destination = `${destination}&host=${host}`;
   }
+  queryParams.set('exitIframe', destination);
 
   throw redirect(`${config.auth.exitIframePath}?${queryParams.toString()}`);
 }