+++ packages/shopify-app-remix/src/server/future/flags.ts
@@ -1,4 +1,7 @@
 // When adding new flags, you should also add them to the `TEST_FUTURE_FLAGS` object in `test-config.ts` to ensure that
+
+import type {ConfigParams, ShopifyRestResources} from '@shopify/shopify-api';
+
 // it doesn't cause regressions.
 export interface FutureFlags {
   /**
@@ -16,6 +19,11 @@ export interface FutureFlags {
   v3_authenticatePublic?: boolean;
 
   /**
+   * When enabled allows you to pass billing plans with line items when creating a new app subscriptions.
+   */
+  v3_lineItemBilling?: boolean;
+
+  /**
    * When enabled, embedded apps will fetch access tokens via [token exchange](https://shopify.dev/docs/apps/auth/get-access-tokens/token-exchange).
    * This assumes the app has scopes declared for [Shopify managing installation](https://shopify.dev/docs/apps/auth/installation#shopify-managed-installation).
    *
@@ -25,6 +33,17 @@ export interface FutureFlags {
    */
   unstable_newEmbeddedAuthStrategy?: boolean;
 }
+
+// When adding new flags, use this format:
+// vX_myFutureFlag: Future extends FutureFlags ? Future['vX_myFutureFlag'] : false;
+export interface ApiFutureFlags<Future extends FutureFlagOptions> {
+  v10_lineItemBilling: Future extends FutureFlags
+    ? Future['v3_lineItemBilling']
+    : false;
+}
+
+export type ApiConfigWithFutureFlags<Future extends FutureFlagOptions> =
+  ConfigParams<ShopifyRestResources, ApiFutureFlags<Future>>;
 
 export type FutureFlagOptions = FutureFlags | undefined;